Added a wrapper function for catching and allowing EC2ResponseErrors containing the allowed text.
Required replacing all the duplicate code with the new function call. Currently, the
boto call args and kwargs are passed in as keyword arguments to the wrapper function, but we
may want to just use a positional kwargs param for everything.

Also, added a constant dict of common allowed EC2ResponseErrors to avoid lots of duplicate
text.

NOTE: There may very well be more places where the allowed EC2ResponseErrors need to be caught.